+++ src/App.test.js
@@ -1,4 +1,4 @@
-import { render } from '@testing-library/react';
+import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import App from './App';
 
 jest.mock('react-uuid', () => () => '0000000-000-0000-000-00000000000');
@@ -11,4 +11,12 @@ jest.mock('../package.json', () => ({
 test('App snapshot', () => {
   const { container } = render(<App />);
   expect(container).toMatchSnapshot();
+});
+
+test('same vessel type', async () => {
+  render(<App />);
+
+  fireEvent.click(screen.getAllByTestId('radio')[5]); // Vessel B, type: BIG_SHIP
+
+  expect(screen.getByTestId('notification')).toMatchSnapshot();
 });
